Deduplication of classes
Moves identical classes to base-module
